RateMe task/navigation  Added the ability to navigate and added authentication state to redux
