+++ App.tsx
@@ -1,23 +1,12 @@
 import React from 'react';
-import { Platform } from 'react-native';
-import { Provider, useSelector } from 'react-redux';
-import { store } from './redux/store/store';
-import {
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  View,
-  Text,
-  RefreshControl,
-  Button
-} from 'react-native';
+import { useSelector } from 'react-redux';
+import { StyleSheet } from 'react-native';
 import SignIn from './components/SignIn';
 import SignUp from './components/SignUp';
-import LoggedIn from './components/LoggedIn';
+import Home from './components/Home';
 import { authSelector } from './redux/reducers/authSlice';
-import {NavigationContainer} from '@react-navigation/native';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 const styles = StyleSheet.create({
   outerContainer: {
@@ -31,36 +20,35 @@ const styles = StyleSheet.create({
     backgroundColor: '#fff',
   },
 });
+
 const Stack = createNativeStackNavigator();
 
-function App() {
+const App = () => {
+  const isLoggedIn = useSelector(authSelector);
 
-   const isLoggedIn = useSelector(authSelector);
+  const [refreshing, setRefreshing] = React.useState(false);
 
-   const [refreshing, setRefreshing] = React.useState(false);
-
-    const onRefresh = React.useCallback(() => {
-      setRefreshing(true);
-      setTimeout(() => {
-        setRefreshing(false);
-      }, 2000);
-    }, []);
+  const onRefresh = React.useCallback(() => {
+    setRefreshing(true);
+    setTimeout(() => {
+      setRefreshing(false);
+    }, 2000);
+  }, []);
 
   return (
-
     <NavigationContainer>
       <Stack.Navigator initialRouteName="SignIn">
-      { !isLoggedIn ? (
-      <>
-      <Stack.Screen name="SignIn" component={SignIn} />
-      <Stack.Screen name="SignUp" component={SignUp} />
-      </> ) : (
-      <Stack.Screen name="LoggedIn" component={LoggedIn} />
-      )}
+        {!isLoggedIn ? (
+          <>
+            <Stack.Screen name="SignIn" component={SignIn} />
+            <Stack.Screen name="SignUp" component={SignUp} />
+          </>
+        ) : (
+          <Stack.Screen name="Home" component={Home} />
+        )}
       </Stack.Navigator>
-
     </NavigationContainer>
   );
-}
+};
 
 export default App;
