+++ index.js
@@ -3,18 +3,18 @@
  */
 
 import React from 'react';
-import {AppRegistry} from 'react-native';
+import { AppRegistry } from 'react-native';
 import App from './App';
-import {name as appName} from './app.json';
+import { name as appName } from './app.json';
 import { Provider } from 'react-redux';
 import { store } from './redux/store/store';
 
 const RootComponent = () => {
-    return (
-        <Provider store={store}>
-            <App />
-        </Provider>
-    );
+  return (
+    <Provider store={store}>
+      <App />
+    </Provider>
+  );
 };
 
 AppRegistry.registerComponent(appName, () => RootComponent);
