+++ redux/reducers/platformSlice.ts
@@ -1,5 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { RootState } from '../store/store';
+import { Platform } from 'react-native';
 
 interface OS {
   os: string;
@@ -11,7 +12,8 @@ const initialState: OS = {
 
 const osSlice = createSlice({
   name: 'OS',
-  initialState
+  initialState,
+  reducers: {},
 });
 
 export const osSelector = (state: RootState) => state.osReducer.os;
